+++ app/components/our-story/index.ts
@@ -1,5 +1,3 @@
 export { OurStoryHero } from "./OurStoryHero";
 export { StorySection } from "./StorySection";
-export { StoryManifesto } from "./StoryManifesto";
-export { StoryRail } from "./StoryRail";
 export { OurStoryCTA } from "./OurStoryCTA";